Reorganizes CI scripts
