+++ package.json
@@ -7,10 +7,10 @@
     "license": "GNU General Public License v3.0",
     "type": "module",
     "scripts": {
-        "install-resources-data": "npx tsx ci/install-resources-data.ts",
+        "install-resources-data": "npx tsx ci/postinstall/install-resources-data.ts",
         "rebuild-sqlite3": "electron-rebuild -f -w sqlite3",
         "postinstall": "npm run install-resources-data && npm run rebuild-sqlite3",
-        "init-app": "npx tsx ci/init-app.ts",
+        "init-app": "npx tsx ci/app/init.ts",
         "compile-typescript": "tsc",
         "compile-react": "webpack --mode development --config webpack.react.config.js",
         "electron": "electron .",
